Use config with necessary data
This commit fixes an error and two warnings during tests.

This config points to a folder containing the relevant data. It
seems that such data was available before but not now in the
directory pointed by the default config. This suggests that someone
changed this crucial 2dii folder. This exposes the need for a more
robust approach for testing this features. Maybe create a test
folder with a name such as 'TEST_quarter_export_date' or similar.